feat(e2e-test-plugins): migrate plugins to use rspack
